Improving the asset metadata editor window
`AssetMetadataEditor` and `MaterialEditor` are now more consistent.
Their use flow is now similar. You can either right click onto an asset
and choose to edit its metadata or edit the material if it is one OR
open the desired window (`AssetMetadataEditor` or `MaterialEditor`) and
then drag and drop an asset to this window.

The `AssetMetadataEditor` is now divided into two sections:
- Settings
- Info

The first one, "Settings", will show every metadata associated to this
asset and allow the user to modify their values.

The second one, "Info", only display information about the asset (File
type, file size, path...)

This commit is WIP, todo:
- [ ] Change the `AssetMetadataEditor` name to better reflect its new content
- [ ] Improve the code
- [ ] Add more info to the asset info section